refactor(omo-senpi): split recall wiring into session readers, drain, and collection
Plan: .omo/plans/memorian-tool-boundary.md

Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)
